+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.86.1] - 2019-10-03
+
 ### Fixed
 
+- Label PropTypes of `Input`, `TextArea` and `Toggle`.
+- Value PropType of `Input`.
 - `Button` hover state in collapsed tertiary buttons
 
 ## [9.86.0] - 2019-10-03